refactor(Lottos): 총 상금액 계산 로직 확장함수 이용해 로직 수정

+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -77,7 +77,7 @@ class LottoController(
         purchasedAmount: LottoPurchaseAmount,
     ) {
         outputView.printLottoStatistics(lottos.ranks)
-        val totalPrizeMoney: Double = lottos.getTotalPrizeMoney()
+        val totalPrizeMoney: Long = lottos.getTotalPrizeMoney()
         val rateOfReturn: Double = lottoCalculator.getRateOfReturn(totalPrizeMoney, purchasedAmount.money)
         outputView.printLottoRateOfReturn(rateOfReturn, lottoCalculator.getIsLossMoney(rateOfReturn))
     }
